+++ extensions/volumesnapshot/tests/src/test/java/io/fabric8/volumesnapshot/test/crud/VolumeSnapshotTest.java
@@ -25,7 +25,6 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @EnableKubernetesMockClient(crud = true)
 class VolumeSnapshotTest {
@@ -96,7 +95,7 @@ class VolumeSnapshotTest {
     assertEquals("true", u1.getMetadata().getLabels().get("updated"));
 
     //Delete
-    assertTrue(client.volumeSnapshots().withName("my-snapshot1").delete().size() == 1);
+    assertEquals(1, client.volumeSnapshots().withName("my-snapshot1").delete().size());
     assertNull(client.volumeSnapshots().withName("my-snapshot1").get());
   }
 }